projects/pythonjobs-v1/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
stripe==12.4.0
jinja2==3.1.6
